refactor: Extract HomeScreen components and helpers
- Extracts UI components from `HomeScreen.kt` into their own composable files:
    - `GarmentSliders.kt`: Manages the horizontal pagers for each clothing slot.
    - `GarmentSlider.kt` & `GarmentItem.kt`: Individual pager item, lock button, and click handling. Includes logic for the "None" headwear option.
    - `HomeScreenTopBar.kt`: The top bar with "Your Outfit" title, shuffle, and favorite buttons.
    - `PiecePickerDialog.kt`: The dialog for selecting a piece from a grid view.
    - `BackgroundContent.kt`: Logic for displaying different screen backgrounds.
- Introduces `HomeScreenState.kt` to manage the screen's state, including pager states, locked pieces, and the current outfit's saved status.
- Moves complex logic and utility functions from `HomeScreen.kt` into `util/HomeScreenHelpers.kt`.
- Refactors `FilteredView.kt` and `OutfitImageGenerator.kt` by extracting smaller, more focused helper functions to improve readability and maintainability.
